+++ src/biz/bokhorst/xprivacy/ActivityShare.java
@@ -690,8 +690,13 @@ public class ActivityShare extends ActivityBase {
 
 						List<Boolean> oldState = PrivacyManager.getRestartStates(uid, restrictionName);
 
-						if (actionId == R.id.rbClear)
+						if (actionId == R.id.rbClear) {
 							PrivacyManager.deleteRestrictions(uid, restrictionName, (restrictionName == null));
+							if (restrictionName == null) {
+								PrivacyManager.deleteUsage(uid);
+								PrivacyManager.deleteSettings(uid);
+							}
+						}
 
 						else if (actionId == R.id.rbRestrict) {
 							PrivacyManager.setRestriction(uid, restrictionName, null, true, false);